requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pyodbc
python-dotenv